+++ src/main/java/racingcar/controller/WebRacingCarController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import racingcar.dto.ResultDTO;
-import racingcar.request.RacingGameRequest;
-import racingcar.response.RacingGameResponse;
+import racingcar.dto.request.RacingGameRequest;
+import racingcar.dto.response.RacingGameResponse;
 import racingcar.exception.CarException;
 import racingcar.exception.CarsException;
 import racingcar.service.RacingCarService;
