+++ src/test/java/racingcar/controller/WebRacingCarControllerTest.java
@@ -9,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import racingcar.request.RacingGameRequest;
+import racingcar.dto.request.RacingGameRequest;
 
 import static org.hamcrest.core.Is.is;
